rename list to nodes
