+++ discovery/discovery.go
@@ -55,6 +55,11 @@ func New(rawurl string, heartbeat int) (DiscoveryService, error) {
 		return nil, err
 	}
 
+	// nodes:port,node2:port => nodes://node1:port,node2:port
+	if url.Scheme == "" {
+		url.Scheme = "nodes"
+	}
+
 	if discovery, exists := discoveries[url.Scheme]; exists {
 		log.Debugf("Initializing %q discovery service with %q", url.Scheme, url.Host+url.Path)
 		err := discovery.Initialize(url.Host+url.Path, heartbeat)
